requirements.txt
certifi==2024.8.30
geographiclib==2.0
pillow==10.4.0
pyproj==3.6.1
